Multiple changes for final lab prep
2 friends fragments created.  one to use the eqation fragment and use it
to search for users in the database.  The other to display current
friends in the button_fragment.

All potential methods created and pre-commented as well.  These methods
include getFriendsEquation() ,findNewFriends(), addFriend(),
viewOPendingRequests()

Three helper methods to load up the fragments when the friends button is
pushed showFriendsSearch, loadFriendsFragments, and showFriends